vagrant: add Vagrantfile with quantal support
Vagrant is a command line tool for working with virtual machines for
reproducible test and development environment. Vagrant is packaged
in Ubuntu and uses VirtualBox as the implementation backend which makes
it highly portable and available, regardless of access to modern
hardware with virtualisation extensions.

The vagrant file added here uses the new cloud images built by
canonical. They come with everything pre-installed and should be a
stable base for testing the relevant releases.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>