vagrant: fix broken/missing plainbox.egg-info in test-in-vagrant.sh
This elusive bug kept the CI loop broken/offline for a couple of days.
There are several earlier attempts to fix this but they were all mislead
into fixing non-issues.

The root issue is that "plainbox/setup.py test" fails iff:
 * plainbox was 'developed' by running plainbox/setup.py develop earlier
 * plainbox/plainbox.egg-info (directory) was removed

When those two things happen in sequence, setuptools is tripping over
a bug where it does not re-generate egg-info and crashes while trying
to open it.

The workaround is to always rebuild egg-info before touching plainbox
tests or running the executable generated by distribute.

Testing instructions (verified on 12.04):

 1) $ pip install distribute==0.6.35
 2) $ sudo python3 plainbox/setup.py develop
 3) $ rm -rf plainbox/plainbox.egg-info
 4) $ plainbox/setup.py test

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>