support: add support for depending on Ubuntu PPAs
This new support code allows any component to express PPA dependencies
(that a particular PPA is enabled in the system). This can be done by
creating requirement/repo-*.txt files that hold the name of the APT
repository to add. The full set of repositories is sorted and
de-duplicated and enabled one-by-one with the help of
'add-apt-repository'

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>