test-in-lxc: Fix file permissions if provisioning fails.
Writing files in the source tree during the provisioning phase is
not recommended, but if it happens, ensure that, after a possible
provisioning failure, permissions are reset to those of the user who
owned the original tree (using test-in-lxc.sh as reference).

Previously we would only fix file permissions after running the tests,
and the provisioning phase was assumed not to leave possibly-undeletable
files in the source directory.